requirements.txt
# Core dependencies
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
python-dotenv==1.0.0

# CAMEL framework and LLM
openai==1.3.9
camel-ai==0.1.1

# OSINT tools
snscrape==0.7.0.20230622
yt-dlp==2023.7.6
python-geocoder==1.38.1
exifread==3.0.0
pillow==10.0.0
opencv-python==4.8.0.76

# Database
sqlalchemy==2.0.20
sqlitedict==2.1.0

# GUI
pyside6==6.5.2

# API integrations
requests==2.31.0
aiohttp==3.8.5 